Tag for release: 20.6.0

+++ install.sh
@@ -14,6 +14,8 @@ MIN_RAM=2400 # MB
 
 SENTRY_CONFIG_PY='sentry/sentry.conf.py'
 SENTRY_CONFIG_YML='sentry/config.yml'
+RELAY_CONFIG_YML='relay/config.yml'
+RELAY_CREDENTIALS_JSON='relay/credentials.json'
 SENTRY_EXTRA_REQUIREMENTS='sentry/requirements.txt'
 
 DID_CLEAN_UP=0
@@ -62,11 +64,15 @@ if [ "$RAM_AVAILABLE_IN_DOCKER" -lt "$MIN_RAM" ]; then
     exit 1
 fi
 
-#SSE4.2 required by Clickhouse (https://clickhouse.yandex/docs/en/operations/requirements/) 
-SUPPORTS_SSE42=$(docker run --rm busybox grep -c sse4_2 /proc/cpuinfo || :);
-if (($SUPPORTS_SSE42 == 0)); then
+#SSE4.2 required by Clickhouse (https://clickhouse.yandex/docs/en/operations/requirements/)
+# On KVM, cpuinfo could falsely not report SSE 4.2 support, so skip the check. https://github.com/ClickHouse/ClickHouse/issues/20#issuecomment-226849297
+IS_KVM=$(docker run --rm busybox grep -c 'Common KVM processor' /proc/cpuinfo || :)
+if (($IS_KVM == 0)); then
+  SUPPORTS_SSE42=$(docker run --rm busybox grep -c sse4_2 /proc/cpuinfo || :)
+  if (($SUPPORTS_SSE42 == 0)); then
     echo "FAIL: The CPU your machine is running on does not support the SSE 4.2 instruction set, which is required for one of the services Sentry uses (Clickhouse). See https://git.io/JvLDt for more info."
     exit 1
+  fi
 fi
 
 # Clean up old stuff and ensure nothing is working while we install/update
@@ -98,22 +104,81 @@ if grep -xq "system.secret-key: '!!changeme!!'" $SENTRY_CONFIG_YML ; then
     echo "Secret key written to $SENTRY_CONFIG_YML"
 fi
 
+replace_tsdb() {
+    if (
+        [ -f "$SENTRY_CONFIG_PY" ] &&
+        ! grep -xq 'SENTRY_TSDB = "sentry.tsdb.redissnuba.RedisSnubaTSDB"' "$SENTRY_CONFIG_PY"
+    ); then
+        tsdb_settings="SENTRY_TSDB = \"sentry.tsdb.redissnuba.RedisSnubaTSDB\"
+
+# Automatic switchover 90 days after $(date). Can be removed afterwards.
+SENTRY_TSDB_OPTIONS = {\"switchover_timestamp\": $(date +%s) + (90 * 24 * 3600)}"
+
+        if grep -q 'SENTRY_TSDB_OPTIONS = ' "$SENTRY_CONFIG_PY"; then
+            echo "Not attempting automatic TSDB migration due to presence of SENTRY_TSDB_OPTIONS"
+        else
+            echo "Attempting to automatically migrate to new TSDB"
+            # Escape newlines for sed
+            tsdb_settings="${tsdb_settings//$'\n'/\\n}"
+            cp "$SENTRY_CONFIG_PY" "$SENTRY_CONFIG_PY.bak"
+            sed -i -e "s/^SENTRY_TSDB = .*$/${tsdb_settings}/g" "$SENTRY_CONFIG_PY" || true
+
+            if grep -xq 'SENTRY_TSDB = "sentry.tsdb.redissnuba.RedisSnubaTSDB"' "$SENTRY_CONFIG_PY"; then
+                echo "Migrated TSDB to Snuba. Old configuration file backed up to $SENTRY_CONFIG_PY.bak"
+                return
+            fi
+
+            echo "Failed to automatically migrate TSDB. Reverting..."
+            mv "$SENTRY_CONFIG_PY.bak" "$SENTRY_CONFIG_PY"
+            echo "$SENTRY_CONFIG_PY restored from backup."
+        fi
+
+        echo "WARN: Your Sentry configuration uses a legacy data store for time-series data. Remove the options SENTRY_TSDB and SENTRY_TSDB_OPTIONS from $SENTRY_CONFIG_PY and add:"
+        echo ""
+        echo "$tsdb_settings"
+        echo ""
+        echo "For more information please refer to https://github.com/getsentry/onpremise/pull/430"
+    fi
+}
+
+replace_tsdb
+
+echo ""
+echo "Fetching and updating Docker images..."
+echo ""
+# We tag locally built images with an '-onpremise-local' suffix. docker-compose pull tries to pull these too and
+# shows a 404 error on the console which is confusing and unnecessary. To overcome this, we add the stderr>stdout
+# redirection below and pass it through grep, ignoring all lines having this '-onpremise-local' suffix.
+$dc pull -q --ignore-pull-failures 2>&1 | grep -v -- -onpremise-local || true
+
+if [ -z "$SENTRY_IMAGE" ]; then
+  docker pull getsentry/sentry:${SENTRY_VERSION:-latest}
+else
+  # We may not have the set image on the repo (local images) so allow fails
+  docker pull $SENTRY_IMAGE || true;
+fi
+
 echo ""
 echo "Building and tagging Docker images..."
 echo ""
 # Build the sentry onpremise image first as it is needed for the cron image
-$dc pull --ignore-pull-failures
-docker pull ${SENTRY_IMAGE:-getsentry/sentry:latest}
 $dc build --force-rm web
 $dc build --force-rm --parallel
 echo ""
 echo "Docker images built."
 
-echo "Bootstrapping Snuba..."
-# `bootstrap` is for fresh installs, and `migrate` is for existing installs
-# Running them both for both cases is harmless so we blindly run them
+
+ZOOKEEPER_LOG_FILE_COUNT=$($dcr zookeeper bash -c 'ls 2>/dev/null -Ubad1 -- /var/lib/zookeeper/log/version-2/* | wc -l | tr -d '[:space:]'')
+ZOOKEEPER_SNAPSHOT_FILE_COUNT=$($dcr zookeeper bash -c 'ls 2>/dev/null -Ubad1 -- /var/lib/zookeeper/data/version-2/* | wc -l | tr -d '[:space:]'')
+# This is a workaround for a ZK upgrade bug: https://issues.apache.org/jira/browse/ZOOKEEPER-3056
+if [ "$ZOOKEEPER_LOG_FILE_COUNT" -gt "0" ] && [ "$ZOOKEEPER_SNAPSHOT_FILE_COUNT" -eq "0" ]; then
+  $dcr -v $(pwd)/zookeeper:/temp zookeeper bash -c 'cp /temp/snapshot.0 /var/lib/zookeeper/data/version-2/snapshot.0'
+  $dc run -d -e ZOOKEEPER_SNAPSHOT_TRUST_EMPTY=true zookeeper
+fi
+
+
+echo "Bootstrapping and migrating Snuba..."
 $dcr snuba-api bootstrap --force
-$dcr snuba-api migrate
 echo ""
 
 # Very naively check whether there's an existing sentry-postgres volume and the PG version in it
@@ -159,6 +224,33 @@ if [ "$SENTRY_DATA_NEEDS_MIGRATION" ]; then
   # The `\"` escape pattern is to make this compatible w/ Git Bash on Windows. See #329.
   $dcr --entrypoint \"/bin/bash\" web -c \
     "mkdir -p /tmp/files; mv /data/* /tmp/files/; mv /tmp/files /data/files; chown -R sentry:sentry /data"
+fi
+
+
+if [ ! -f "$RELAY_CREDENTIALS_JSON" ]; then
+  echo ""
+  echo "Generating Relay credentials..."
+
+  # We need the ugly hack below as `relay generate credentials` tries to read the config and the credentials
+  # even with the `--stdout` and `--overwrite` flags and then errors out when the credentials file exists but
+  # not valid JSON. We hit this case as we redirect output to the same config folder, creating an empty
+  # credentials file before relay runs.
+  $dcr --no-deps -v $(pwd)/$RELAY_CONFIG_YML:/tmp/config.yml relay --config /tmp credentials generate --stdout > "$RELAY_CREDENTIALS_JSON"
+  echo "Relay credentials written to $RELAY_CREDENTIALS_JSON"
+fi
+
+RELAY_CREDENTIALS=$(sed -n 's/^.*"public_key"[[:space:]]*:[[:space:]]*"\([a-zA-Z0-9_-]\{1,\}\)".*$/\1/p' "$RELAY_CREDENTIALS_JSON")
+if [ -z "$RELAY_CREDENTIALS" ]; then
+  >&2 echo "FAIL: Cannot read credentials back from $RELAY_CREDENTIALS_JSON."
+  >&2 echo "      Please ensure this file is readable and contains valid credentials."
+  >&2 echo ""
+  exit 1
+fi
+
+if ! grep -q "\"$RELAY_CREDENTIALS\"" "$SENTRY_CONFIG_PY"; then
+  echo "SENTRY_RELAY_WHITELIST_PK = (SENTRY_RELAY_WHITELIST_PK or []) + ([\"$RELAY_CREDENTIALS\"])" >> "$SENTRY_CONFIG_PY"
+  echo "Relay public key written to $SENTRY_CONFIG_PY"
+  echo ""
 fi
 
 cleanup
